requirements.txt
flask
requests
python-dotenv
pytz
gunicorn
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib

